pyproject.toml
[project]
name = "django-starter-template"
version = "0.1.0"
description = ""
authors = [
    {name = "prashunchitkr",email = "prashunchi@gmail.com"}
]
readme = "README.md"
requires-python = ">3.13,<4.0"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "django-configurations[cache] (>=2.5.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"

